Refactored file names of UI modules
